shop/requirements.txt
django==3.2.3
ipython==7.27.0
flake8==3.9.2
flake8-django==1.1.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
django-celery-results==2.2.0
requests==2.26.0
django-bulk-update-or-create==0.3.0
django-celery-beat==2.2.1
django-bootstrap4==3.0.1
Pillow==8.3.2
psycopg2-binary==2.9.1
amqp==5.2.0
asgiref==3.8.1
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==3.2.3
django-bootstrap4==3.0.1
django-bulk-update-or-create==0.3.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.3
flake8==3.9.2
flake8-builtins==1.5.3
flake8-django==1.1.2
flake8-import-order==0.18.1
flake8-print==4.0.0
idna==3.7
ipython==7.27.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
mccabe==0.6.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.17.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.26.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
